requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
deep-phonemizer==0.0.19
exceptiongroup==1.2.2
executing==2.2.0
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
grpcio==1.70.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip-autoremove==0.10.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cord==2.6.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3
